requirements.txt
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
fastapi==0.115.0
uvicorn==0.30.6
PyYAML==6.0.2
joblib==1.4.2
tensorflow==2.15.0
pydantic==2.8.2
